requirements.txt
fastapi==0.75.0
pandas==1.3.4
python-dotenv==0.19.2
python_dateutil==2.8.2
SQLAlchemy==1.4.32
uvicorn==0.17.4
psycopg2-binary
pydantic==1.8.2






